chore: add license and move to utils folder
